Add permission middleware

+++ src/items/items.router.ts
@@ -7,6 +7,8 @@ import * as ItemService from './items.service';
 import { BaseItem, Item } from './item.interface';
 
 import { checkJwt } from '../middleware/authz.middleware';
+import { checkPermissions } from '../middleware/permission.middleware';
+import { ItemPermission } from './item-permission';
 
 /**
  * Router definition
@@ -48,52 +50,63 @@ itemsRouter.get('/:id', async (req: Request, res: Response) => {
   }
 });
 
+// Mount authorization middleware
+
 itemsRouter.use(checkJwt);
 
 // POST items
 
-itemsRouter.post('/',  async (req: Request, res: Response) => {
-  try {
-    const item: BaseItem = req.body;
-    const newItem = await ItemService.create(item);
+itemsRouter.post(
+  '/',
+  checkPermissions(ItemPermission.CreateItems),
+  async (req: Request, res: Response) => {
+    try {
+      const item: BaseItem = req.body;
+      const newItem = await ItemService.create(item);
 
-    res.status(201).json(newItem);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+      res.status(201).json(newItem);
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
 });
 
 // PUT items/:id
 
-itemsRouter.put('/:id', async (req: Request, res: Response) => {
-  const id: number = parseInt(req.params.id, 10);
+itemsRouter.put(
+  '/:id',
+  checkPermissions(ItemPermission.CreateItems),
+  async (req: Request, res: Response) => {
+    const id: number = parseInt(req.params.id, 10);
 
-  try {
-    const itemUpdate: Item = req.body;
-    const existingItem: Item = await ItemService.find(id);
+    try {
+      const itemUpdate: Item = req.body;
+      const existingItem: Item = await ItemService.find(id);
 
-    if (existingItem) {
-      const updatedItem = await ItemService.update(id, itemUpdate);
-      return res.status(200).json(updatedItem);
+      if (existingItem) {
+        const updatedItem = await ItemService.update(id, itemUpdate);
+        return res.status(200).json(updatedItem);
+      }
+
+      const newItem = await ItemService.create(itemUpdate);
+
+      res.status(201).json(newItem);
+    } catch (e) {
+      res.status(500).send(e.message);
     }
-
-    const newItem = await ItemService.create(itemUpdate);
-
-    res.status(201).json(newItem);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
 });
 
 // DELETE items/:id
 
-itemsRouter.delete('/:id', async (req: Request, res: Response) => {
-  try {
-    const id: number = parseInt(req.params.id, 10);
-    await ItemService.remove(id);
-    
-    res.sendStatus(204);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+itemsRouter.delete(
+  '/:id',
+  checkPermissions(ItemPermission.CreateItems),
+  async (req: Request, res: Response) => {
+    try {
+      const id: number = parseInt(req.params.id, 10);
+      await ItemService.remove(id);
+      
+      res.sendStatus(204);
+    } catch (e) {
+      res.status(500).send(e.message);
+    }
 });
